[sparkle] refactor: rename puzzle spinner
